Writing logfiles to the workspace.
When running scenario tasks in parallel (which is convenient on
Sherlock), logfiles are by default written to $HOME, which is an NFS
partition only 15GB in size.  Plus, log messages from each run are
interleaved, making it difficult to debug issues.  Writing the logs to
the workspace allows for clearer association of which logfiles belong
with which scenario.